requirements.txt
yt-dlp
webvtt-py
feedparser
openai
python-dotenv
requests
